backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
facebook_business>=21.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
python-multipart>=0.0.9
email-validator>=2.0.0
httpx>=0.24.0
